requirements.txt
anyio==3.6.2
appnope==0.1.3
asttokens==2.2.1
async-timeout==5.0.1
asyncpg==0.30.0
backcall==0.2.0
bcrypt==4.0.1
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
cryptography==41.0.1
cssselect2==0.7.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.7.0
ecdsa==0.18.0
email_validator==2.2.0
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.95.0
fastapi-utils==0.2.1
fonttools==4.42.1
Glances==3.3.1.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpie==3.2.1
httptools==0.5.0
httpx==0.28.1
idna==3.4
ipykernel==6.23.3
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.6.3
jupyter_core==5.8.1
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
mysql-connector-python-rf==2.2.2
nest-asyncio==1.5.6
numpy==1.25.0
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
pango==0.0.1
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.8.0
prompt-toolkit==3.0.38
psutil==5.9.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
pydyf==0.7.0
Pygments==2.14.0
PyMySQL==1.0.3
pyphen==0.14.0
pypng==0.20220715.0
PyQRCode==1.2.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.1.0
requests==2.28.2
requests-toolbelt==0.10.1
rich==13.3.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.47
stack-data==0.6.2
starlette==0.26.1
tinycss2==1.2.1
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.18.1
wcwidth==0.2.6
weasyprint==59.0
webencodings==0.5.1
websockets==10.4
zopfli==0.2.2
